Refactor to fix API for unit tests and utilities
