+++ sns_ik_lib/test/sns_ik_vel_test.cpp
@@ -1,5 +1,5 @@
-/*! \file sns_ik_test.cpp
- * \brief Unit Test: sns_ik_math_utils
+/*! \file sns_ik_vel_test.cpp
+ * \brief Unit Test: sns_ik velocity solver
  * \author Matthew Kelly
  */
 /*
@@ -28,8 +28,8 @@
 #include <ros/duration.h>
 #include <ros/time.h>
 
-#include <sns_ik/rng_utilities.hpp>
-#include <sns_ik/robot_model.hpp>
+#include "rng_utilities.hpp"
+#include "robot_model.hpp"
 #include <sns_ik/sns_ik.hpp>
 
 /*
@@ -225,6 +225,7 @@ void runGeneralVelIkTest(int seed, KDL::ChainFkSolverVel_recursive& fwdKin, IkSo
 /*************************************************************************************************/
 
 void runSnsVelkTest(int seed, sns_ik::VelocitySolveType solverType) {
+
   // Create a sawyer model:
   std::vector<std::string> jointNames;
   KDL::Chain sawyerChain = sns_ik::robot_model::getSawyerKdlChain(&jointNames);
